+++ examples/ios/swift/ListSwiftUI/Views/ContentView.swift
@@ -44,7 +44,7 @@ class Reminder: EmbeddedObject, ObjectKeyIdentifiable {
 class ReminderList: Object, ObjectKeyIdentifiable {
     @Persisted var name = "New List"
     @Persisted var icon: String = "list.bullet"
-    @persisted var reminders: RealmSwift.List<Reminder>
+    @Persisted var reminders: RealmSwift.List<Reminder>
 }
 
 struct FocusableTextField: UIViewRepresentable {